requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.3.0
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.8.2
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
comm==0.1.4
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
email_validator==2.1.1
et-xmlfile==1.1.0
executing==1.2.0
faiss-cpu==1.9.0.post1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.18.0
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
Flask-WTF==1.2.1
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.78.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.15.0
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-status==1.62.3
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.26.1
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.7
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-google-genai==2.0.9
langchain-groq==0.2.3
langchain-text-splitters==0.3.5
langcodes==3.4.0
langsmith==0.3.0
language_data==1.2.0
llvmlite==0.43.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.25.1
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.4.2
nltk==3.8.1
notebook==7.0.2
notebook_shim==0.2.3
numba==0.60.0
numpy==1.26.4
openpyxl==3.1.5
orjson==3.10.3
overrides==7.4.0
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.3.0
platformdirs==3.10.0
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
psutil==5.9.5
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pygame==2.5.2
Pygments==2.16.1
PyMuPDF==1.24.12
pyod==2.0.2
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.2
regex==2024.5.15
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.9.2
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.2
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.4.8
stack-data==0.6.2
starlette==0.37.2
streamlit==1.35.0
sympy==1.13.1
tenacity==8.3.0
terminado==0.17.1
thinc==8.2.5
threadpoolctl==3.5.0
tinycss2==1.2.1
tokenizers==0.20.1
toml==0.10.2
toolz==0.12.1
torch==2.5.0
tornado==6.3.3
tqdm==4.66.4
traitlets==5.9.0
transformers==4.46.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.29.0
wasabi==1.1.3
watchdog==4.0.1
watchfiles==0.21.0
wcwidth==0.2.6
weasel==0.4.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.8
wrapt==1.16.0
WTForms==3.1.2
yarl==1.18.3
youtube-transcript-api==0.6.3
zstandard==0.23.0
